backend/approval-request-service/requirements.txt
fastapi
uvicorn[standard]

motor
httpx
grpcio
protobuf
aio-pika>=9.3.0,<10